utils unit tests passed
