+++ src/sidebar/components/tools/print/printRequest/__tests__/request.js
@@ -1,0 +1,14 @@
+const http = require('http');
+
+export default function request(url) {
+  return new Promise(resolve => {
+    // This is an example of an http request, for example to fetch
+    // user data from an API.
+    // This module is being mocked in __mocks__/request.js
+    http.get({path: url}, response => {
+      let data = '';
+      response.on('data', _data => (data += _data));
+      response.on('end', () => resolve(data));
+    });
+  });
+}
