+++ src/sidebar/components/tools/print/printRequest/__tests__/testPrintOptions.json
@@ -1,0 +1,16 @@
+{
+    "mapTitle": "County of Simcoe - Web Map",
+    "printSizeSelectedOption": {
+        "value": "8X11 Portrait",
+        "label": "8X11 Portrait (Letter)"
+    },
+    "printFormatSelectedOption": {
+        "value": "PDF",
+        "label": "PDF"
+    },
+    "forceScale": 577791,
+    "mapScaleOption": "forceScale",
+    "mapOnlyHeight": 855,
+    "mapOnlyWidth": 647,
+    "isPrinting": false
+}
